`defmt-parser`: Move tests to separate file
